Run all commands through common process pool
