+++ clowder/clowder/clowder_controller.py
@@ -2,12 +2,9 @@
 
 from __future__ import print_function
 
-import multiprocessing as mp
 import os
-import signal
 import sys
 
-import psutil
 from termcolor import cprint
 
 import clowder.clowder_yaml as clowder_yaml
@@ -15,7 +12,7 @@ import clowder.util.formatting as fmt
 from clowder.error.clowder_error import ClowderError
 from clowder.model.group import Group
 from clowder.model.source import Source
-from clowder.util.progress import Progress
+from clowder.util.process_pool import POOL, PROGRESS, RESULTS
 
 
 def herd(project, branch, tag, depth, rebase):
@@ -36,36 +33,6 @@ def run(project, command, ignore_errors):
 def sync(project, rebase):
     """Sync fork project with upstream"""
     project.sync(rebase, parallel=True)
-
-
-PARENT_ID = os.getpid()
-
-
-def worker_init():
-    """
-    Process pool terminator
-    Adapted from https://stackoverflow.com/a/45259908
-    """
-    def sig_int(signal_num, frame):
-        """Signal handler"""
-        del signal_num, frame
-        # print('signal: %s' % signal_num)
-        parent = psutil.Process(PARENT_ID)
-        for child in parent.children(recursive=True):
-            if child.pid != os.getpid():
-                # print("killing child: %s" % child.pid)
-                child.terminate()
-        # print("killing parent: %s" % parent_id)
-        parent.terminate()
-        # print("suicide: %s" % os.getpid())
-        psutil.Process(os.getpid()).terminate()
-        print('\n\n')
-    signal.signal(signal.SIGINT, sig_int)
-
-
-RESULTS = []
-POOL = mp.Pool(initializer=worker_init)
-PROGRESS = Progress()
 
 
 class ClowderController(object):
